+++ NEWS.md
@@ -4,6 +4,17 @@
   separate tidy-select from data-masking. Use `any_of()` or `all_of()` instead
   (#169).
 
+* New `eval_relocate()` for moving a selection. This powers `dplyr::relocate()`
+  (#232).
+
+* `eval_select()` with `include` now preserves the order of the variables
+  if they're present in the selection (#224).
+
+* `all_of()` now returns an integer vector to make it easier to
+  combine in functions (#270, #294).
+
+* `all_of()` now fails even when `strict` is set to `FALSE`.
+
 * `peek_var()` error now generates hyperlink to docs with recent RStudio (#289).
 
 * `all_of()` (like `any_of()`) now fails when used outside of a